third_party/blink/web_tests/platform/generic/accessibility/input-type-password-value-and-selection-expected.txt

This tests that AXValue, caret position and text selection are reported correctly for password fields.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 0
PASS axPassword.selectionFocusOffset is 0
PASS axPassword.stringValue is "AXValue: ••••••"
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 0
PASS axPassword.selectionFocusOffset is 0
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 0
PASS axPassword.selectionFocusOffset is 1
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 0
PASS axPassword.selectionFocusOffset is 2
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 0
PASS axPassword.selectionFocusOffset is 3
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 0
PASS axPassword.selectionFocusOffset is 4
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 0
PASS axPassword.selectionFocusOffset is 5
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 1
PASS axPassword.selectionFocusOffset is 1
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 1
PASS axPassword.selectionFocusOffset is 2
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 1
PASS axPassword.selectionFocusOffset is 3
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 1
PASS axPassword.selectionFocusOffset is 4
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 1
PASS axPassword.selectionFocusOffset is 5
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 2
PASS axPassword.selectionFocusOffset is 2
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 2
PASS axPassword.selectionFocusOffset is 3
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 2
PASS axPassword.selectionFocusOffset is 4
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 2
PASS axPassword.selectionFocusOffset is 5
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 3
PASS axPassword.selectionFocusOffset is 3
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 3
PASS axPassword.selectionFocusOffset is 4
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 3
PASS axPassword.selectionFocusOffset is 5
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 4
PASS axPassword.selectionFocusOffset is 4
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 4
PASS axPassword.selectionFocusOffset is 5
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 5
PASS axPassword.selectionFocusOffset is 5
PASS axPassword is axPassword.selectionAnchorObject
PASS axPassword is axPassword.selectionFocusObject
PASS axPassword.selectionAnchorOffset is 0
PASS axPassword.selectionFocusOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
